refactor(backend): use standard uuid.UUID and centralize regex patterns
Replaced all custom UUID wrappers with github.com/google/uuid.UUID (stored as VARCHAR(36) in MySQL), updated models, services, repositories, handlers, tests, and migrations accordingly. Removed internal/utils/uuid.go. Moved all symbol and currency regex into internal/utils/regex.go and refactored handlers to use these centralized patterns
