+++ backend/internal/utils/index.go
@@ -1,7 +1,6 @@
 package utils
 
 import (
-	"github.com/google/uuid"
 	"github.com/transaction-tracker/backend/internal/types"
 )
 
@@ -17,9 +16,4 @@ func TradeTypeFromString(s string) (types.TradeType, bool) {
 	default:
 		return "", false
 	}
-}
-
-// GenerateUUID generates a new UUID string (v4)
-func GenerateUUID() string {
-	return uuid.New().String()
 }
